Remove legacy activities
