+++ pretixprint/app/src/main/java/eu/pretix/pretixprint/ui/Settings.kt
@@ -32,7 +32,7 @@ class SettingsFragment : PreferenceFragment() {
         for (type in types) {
             findPreference("hardware_${type}printer_find").setOnPreferenceClickListener {
                 val intent = Intent(activity, PrinterSetupActivity::class.java)
-                intent.putExtra(FindPrinterActivity.EXTRA_TYPE, type)
+                intent.putExtra(PrinterSetupActivity.EXTRA_USECASE, type)
                 activity.startActivity(intent)
                 return@setOnPreferenceClickListener true
             }
